pyproject.toml
[project]
name = "mxbi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "HuYang", email = "hy.ccccr@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "gpiozero>=2.0.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.11.7",
    "pyserial>=3.5",
    "rich>=14.1.0",
    "varname>=0.15.0",
]

[project.scripts]
mxbi = "mxbi:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
